requirements.txt

django>=5.0.0
requests
djangorestframework